qemu-metadata-api: allow building without cgo dependencies for linting
