+++ hack/qemu-metadata-api/server/server_test.go
@@ -1,3 +1,5 @@
+//go:build cgo
+
 /*
 Copyright (c) Edgeless Systems GmbH
