+++ hack/qemu-metadata-api/virtwrapper/BUILD.bazel
@@ -2,7 +2,11 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 
 go_library(
     name = "virtwrapper",
-    srcs = ["virtwrapper.go"],
+    srcs = [
+        "virtwrapper.go",
+        "virtwrapper_cgo.go",
+        "virtwrapper_cross.go",
+    ],
     importpath = "github.com/edgelesssys/constellation/v2/hack/qemu-metadata-api/virtwrapper",
     visibility = ["//visibility:public"],
     deps = ["@org_libvirt_go_libvirt//:libvirt"],
